udemy/crewai_crie_multi_agentes_de_ia_do_zero ao_avancado/estudo_caso_analise_de_mercado/requirements.txt
crewai
crewai_tools
langchain_community
ipykernel
python-dotenv
jupyter
pdfkit
markdown 